edge/ai/requirements.txt
grpcio==1.70.0
grpcio-tools==1.70.0
openai==1.64.0